updated documentation

+++ ov_core/src/sim/Simulator.h
@@ -27,7 +27,7 @@ namespace ov_core {
     /**
      * @brief Master simulator class that will create artificial measurements for our visual-inertial algorithms.
      *
-     * Given a trajectory this will generate a SE(3) bsline for that trajectory.
+     * Given a trajectory this will generate a SE(3) BsplineSE3 for that trajectory.
      * This allows us to get the inertial measurement information at each timestep during this trajectory.
      * After creating the bspline we will generate an environmental feature map which will be used as our feature measurements.
      * This map will be projected into the frame at each timestep to get our "raw" uv measurements.
@@ -192,10 +192,10 @@ namespace ov_core {
         int max_cameras;
 
         /// Frequency of our camera sensors
-        int freq_cam;
+        double freq_cam;
 
         /// Frequency of our imu sensor
-        int freq_imu;
+        double freq_imu;
 
         /// Our running acceleration bias
         Eigen::Vector3d true_bias_accel = Eigen::Vector3d::Zero();
@@ -215,7 +215,7 @@ namespace ov_core {
         /// Gravity in the global frame
         Eigen::Vector3d gravity;
 
-        /// Timeoffset between camera and imu (t_imu = t_cam + t_off)
+        /// Timeoffset between camera and imu (t_cam = t_imu - t_off)
         double calib_camimu_dt;
 
         // Camera intrinsics that we will load in
